API squeaky clean :)
