+++ CategoriesAPI/CategoriesAPI/CategoriesApi.xml
@@ -4,10 +4,75 @@
         <name>CategoriesAPI</name>
     </assembly>
     <members>
-        <member name="T:CategoriesAPI.Controllers.CategoriesController">
+        <member name="M:CategoriesAPI.Controllers.CategoryDBContextController`2.Get">
             <summary>
-            Categories Controller class
+            Get a list of all categories
             </summary>
+            <param name="entity"></param>
+            <returns></returns>
+        </member>
+        <member name="M:CategoriesAPI.Controllers.CategoryDBContextController`2.Get(System.Int32)">
+            <summary>
+            Get category by id
+            </summary>
+            <param name="entity"></param>
+            <returns></returns>
+        </member>
+        <member name="M:CategoriesAPI.Controllers.CategoryDBContextController`2.Put(System.Int32,`0)">
+            <summary>
+            Update a category
+            </summary>
+            <param name="entity"></param>
+            <returns></returns>
+        </member>
+        <member name="M:CategoriesAPI.Controllers.CategoryDBContextController`2.Post(`0)">
+            <summary>
+            Create a category
+            </summary>
+            <param name="entity"></param>
+            <returns></returns>
+        </member>
+        <member name="M:CategoriesAPI.Controllers.CategoryDBContextController`2.Delete(System.Int32)">
+            <summary>
+            Delete a category
+            </summary>
+            <param name="entity"></param>
+            <returns></returns>
+        </member>
+        <member name="M:CategoriesAPI.Data.EFCore.CategoryBaseRepository`2.AddCategory(`0)">
+            <summary>
+            Add a category
+            </summary>
+            <param name="entity"></param>
+            <returns></returns>
+        </member>
+        <member name="M:CategoriesAPI.Data.EFCore.CategoryBaseRepository`2.DeleteCategory(System.Int32)">
+            <summary>
+            Delete a category
+            </summary>
+            <param name="entity"></param>
+            <returns></returns>
+        </member>
+        <member name="M:CategoriesAPI.Data.EFCore.CategoryBaseRepository`2.GetAllCategories">
+            <summary>
+            Get a list of all categories
+            </summary>
+            <param name="entity"></param>
+            <returns></returns>
+        </member>
+        <member name="M:CategoriesAPI.Data.EFCore.CategoryBaseRepository`2.GetCategoryById(System.Int32)">
+            <summary>
+            Get category by id
+            </summary>
+            <param name="entity"></param>
+            <returns></returns>
+        </member>
+        <member name="M:CategoriesAPI.Data.EFCore.CategoryBaseRepository`2.UpdateCategory(`0)">
+            <summary>
+            Update an existing category
+            </summary>
+            <param name="entity"></param>
+            <returns></returns>
         </member>
         <member name="T:CategoriesAPI.Data.ICategoryRepository`1">
             <summary>
