+++ CategoriesAPI/CategoriesAPI/Data/EFCore/CategoryBaseRepository.cs
@@ -1,28 +1,35 @@
 ﻿using Microsoft.EntityFrameworkCore;
-using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
 namespace CategoriesAPI.Data.EFCore
 {
-    public abstract class EfCoreRepository<TEntity, TContext> : ICategoryRepository<TEntity>
-        where TEntity : class, IEntity
+    public abstract class CategoryBaseRepository<TEntity, TContext> : ICategoryRepository<TEntity>
+        where TEntity : class, ICategoryEntity
         where TContext : DbContext
     {
         private readonly TContext context;
 
-        public EfCoreRepository(TContext context)
+        public CategoryBaseRepository(TContext context)
         {
             this.context = context;
         }
-
+        /// <summary>
+        /// Add a category
+        /// </summary>
+        /// <param name="entity"></param>
+        /// <returns></returns>
         public async Task<TEntity> AddCategory(TEntity entity)
         {
             context.Set<TEntity>().Add(entity);
             await context.SaveChangesAsync();
             return entity;
         }
-
+        /// <summary>
+        /// Delete a category
+        /// </summary>
+        /// <param name="entity"></param>
+        /// <returns></returns>
         public async Task<TEntity> DeleteCategory(int id)
         {
             var entity = await context.Set<TEntity>().FindAsync(id);
@@ -36,22 +43,31 @@ namespace CategoriesAPI.Data.EFCore
 
             return entity;
         }
-
+        /// <summary>
+        /// Get a list of all categories
+        /// </summary>
+        /// <param name="entity"></param>
+        /// <returns></returns>
         public async Task<List<TEntity>> GetAllCategories()
         {
             var entity = await context.Set<TEntity>().ToListAsync();
-
             return entity;
-            //throw new NotImplementedException();
-            //TODO
-            // DONT FORGET TODO!!!!
-        }
 
+        }
+        /// <summary>
+        /// Get category by id
+        /// </summary>
+        /// <param name="entity"></param>
+        /// <returns></returns>
         public async Task<TEntity> GetCategoryById(int id)
         {
             return await context.Set<TEntity>().FindAsync(id);
         }
-
+        /// <summary>
+        /// Update an existing category
+        /// </summary>
+        /// <param name="entity"></param>
+        /// <returns></returns>
         public async Task<TEntity> UpdateCategory(TEntity entity)
         {
             context.Entry(entity).State = EntityState.Modified;
