+++ CategoriesAPI/CategoriesAPI/Data/EFCore/CategoryRepository.cs
@@ -1,0 +1,12 @@
+﻿using CategoriesAPI.Models;
+
+namespace CategoriesAPI.Data.EFCore
+{
+    public class CategoryRepository : CategoryBaseRepository<Categories,JeopardyDbContext>
+    {
+        public CategoryRepository(JeopardyDbContext context) : base(context)
+        {
+            
+        }
+    }
+}
