--- CategoriesAPI/CategoriesAPI/Data/EFCore/EfCoreCategoryRepository.cs
@@ -1,12 +1,0 @@
-﻿using CategoriesAPI.Models;
-
-namespace CategoriesAPI.Data.EFCore
-{
-    public class EfCoreCategoryRepository : EfCoreRepository<Categories, AMCDbContext>
-    {
-        public EfCoreCategoryRepository(AMCDbContext context) : base(context)
-        {
-            
-        }
-    }
-}
