+++ CategoriesAPI/CategoriesAPI/Data/ICategoryEntity.cs
@@ -5,8 +5,11 @@ using System.Threading.Tasks;
 
 namespace CategoriesAPI.Data
 {
-    public interface IEntity
+    public interface ICategoryEntity
     {
         int Id { get; set; }
+        string CategoryName_En { get; set; }
+
+        string CategoryName_Fr { get; set; }
     }
 }
