+++ CategoriesAPI/CategoriesAPI/Data/ICategoryRepository.cs
@@ -1,35 +1,34 @@
 ﻿using System.Collections.Generic;
 using System.Threading.Tasks;
-using CategoriesAPI.Data;
 
 namespace CategoriesAPI.Data
 {
     /// <summary>
     /// Interface for Category Repository
     /// </summary>
-    public interface ICategoryRepository<T> where T : class, IEntity
+    public interface ICategoryRepository<T> where T : class, ICategoryEntity
     {    
         /// <summary>
-        /// 
+        /// Get a list of all categories
         /// </summary>
         /// <returns></returns>
         Task<List<T>> GetAllCategories();
 
         /// <summary>
-        /// 
+        /// Get category by id
         /// </summary>
         /// <param name="id"></param>
         /// <returns></returns>
         Task<T> GetCategoryById(int id);        
         
         /// <summary>
-        /// 
+        /// Add a category
         /// </summary>
         /// <param name="entity"></param>
         Task<T> AddCategory(T entity);
 
         /// <summary>
-        /// 
+        /// Update a category
         /// </summary>
         /// <param name="entityToUpdate"></param>
         /// <param name="entity"></param>
@@ -37,7 +36,7 @@ namespace CategoriesAPI.Data
         Task<T> UpdateCategory(T entity);
 
         /// <summary>
-        /// 
+        /// Delete a category
         /// </summary>
         /// <param name="entity"></param>
         Task<T> DeleteCategory(int id);
