+++ CategoriesAPI/CategoriesAPI/Program.cs
@@ -1,12 +1,7 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Hosting;
-using Microsoft.Extensions.Logging;
 using Serilog;
+using System;
 
 namespace CategoriesAPI
 {
@@ -41,6 +36,8 @@ namespace CategoriesAPI
             {
                 Log.CloseAndFlush();
             }
+
+            
         }
 
         public static IHostBuilder CreateHostBuilder(string[] args) =>
