--- CategoriesAPI/CategoriesAPI/ResultSet/CategoryResultSet.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace CategoriesAPI.ResultSet
-{
-    public class CategoryResultSet
-    {
-        public int Id { get; set; }
-        public string Name { get; set; }
-        public bool isActive { get; set; }
-    }
-}
